Rename config-service.js -> config.service.js
