+++ admin/controllers/questions.js
@@ -5,7 +5,7 @@ const uuidv4 = require('uuid/v4')
 const CheckForm = require('../models/check-form')
 const Pupil = require('../models/pupil')
 const School = require('../models/school')
-const configService = require('../services/config-service')
+const configService = require('../services/config.service')
 const jwtService = require('../services/jwt.service')
 
 /**
