+++ admin/spec/routes/questions.spec.js
@@ -9,7 +9,7 @@ const checkFormMock = require('../mocks/checkform')
 const Pupil = require('../../models/pupil')
 const School = require('../../models/school')
 const CheckForm = mongoose.model('Dummy', new Schema({name: {type: String}}))
-const configService = require('../../services/config-service')
+const configService = require('../../services/config.service')
 const httpMocks = require('node-mocks-http')
 const sinon = require('sinon')
 const jwtService = require('../../services/jwt.service')
